+++ support/globals/variables.go
@@ -21,15 +21,18 @@ var Config, AccessData *ini.File
 
 // logFiles
 var SevereFailureThreshold, SensorManagerLog, GateManagerLog, SensorDBLog, EntryManagerLog,
-	SpaceManagerLog, DBSLog, AvgsManagerLog, ApiManager, ExportManager int
+	SpaceManagerLog, DBSLog, AvgsManagerLog, ApiManagerLog, ExportManagerLog int
 
 // Parameters configurable via ini files
 //noinspection GoExportedOwnDeclaration
 var DebugActive, CRCused, SensorEEPROMResetEnabled, CRCMaliciousCount, EnforceStrict, AsymmetricNull,
-	SaveState, Shadowing, AcceptNegatives, ResetCloseTCP, DisableDatabase, DisableCORS, EchoMode bool
+	SaveState, Shadowing, AcceptNegatives, ResetCloseTCP, DisableDatabase, DisableCORS, EchoMode, ExportEnabled,
+	ExportAsync bool
+
 var ChannellingLength, SettleTime, SensorTimeout, TCPdeadline, MaliciousTimeout, MaliciousTriesIP, ServerTimeout,
 	MaliciousTriesMac, MalicioudMode, FailureThreshold, MaximumInvalidIDInternal, ZombieTimeout,
 	RepetitiveTimeout, SecurityLength, SensorEEPROMResetDelay, SensorEEPROMResetStep,
 	AsymmetryMax, AsymmetryIter, ResetPeriod, AsymmetryReset, MaxStateAge int
+
 var APIport, TCPport, DiskCachePath, SensorSettingsFile, ResetSlot, DBpath, DBUser, DBUserPassword,
 	ExportActualCommand, ExportActualArgument, ExportReferenceCommand, ExportReferenceArgument string