+++ webService/variables.go
@@ -1,0 +1,8 @@
+package webService
+
+const (
+	settleTime = 1
+)
+
+var webLog int
+var apiLocation, webServicePort, apiPort string