+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 ## [2.0.0]
 ### This new branch is a complete re-write of the server with no guarantee for back compatibility 
 
+### Added
+ - Possibility to export data via external script  
+ - Option to remove all logs  
+
 ### Changed
  - Server architecture has been changed  
  - Database is no longer based on MongoDB and not badger  
  - Malicious attack check has been simplified, more complex checks will be added later    
  - Several bugs have been removed  
+ - Web Service removed, will be completely redesigned